Git init, DurCal inserts ok
This project takes in a list of tasks: "date~description" and outputs the
duration from the current date. Other functionality (culling obsoltete,
restricting shown) will follow. This is yet another program I've wanted to
make for myself since I began.

There may be a more appropriate DS than a list for the tasks, but probably
not given the insertions and so on. Ultimately, my sensitivity and use
case will signal whether a list is robust enough or if moving to a tree is
more appropriate.

Insertion seems to work alright with the lightweight task file & explicit
insertion. Next involves the reason d' etre: outputting the durations. In
fact, I think I was about to surrender the wrong list. Aha, Tasks is
incomplete. It doesn't just need the end date, it needs the duration()
left for that task.